Highlighted color scale.

This scale blends the hue_pal and grey_pal palettes from the scales package. The "highlighted" elements, selected by position, use the hue palette; the other elements use the grey pallete.

scale_colour_focus(pos, ..., h = c(0, 360) + 15, c = 100, l = 65,
  h.start = 0, direction = 1, start = 0.2, end = 0.8,
  na.value = "red")

Arguments

pos
One or more positions in color scale to focus on

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

h
range of hues to use, in [0, 360]

c
chroma (intensity of colour), maximum value varies depending on combination of hue and luminance.

l
luminance (lightness), in [0, 100]

h.start
hue to start at

direction
direction to travel around the colour wheel, 1 = clockwise, -1 = counter-clockwise

start
gray value at low end of palette

end
gray value at high end of palette

na.value
Colour to use for missing values